pyproject.toml
[project]
name = "rely"
version = "0.1.0"
description = "Rely is a tool for vetting dependencies on GitHub."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.0",
    "fastapi[standard]>=0.119.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.0",
    "streamlit>=1.50.0",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "datamodel-code-generator>=0.35.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.22",
    "mkdocstrings[python]>=0.30.1",
    "mypy>=1.18.2",
    "rich>=14.2.0",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
